+++ R/formulas.R
@@ -32,6 +32,12 @@ drop_special <- function(terms, special) {
   }
 }
 
+has_special_strata <- function(formula) {
+  !is.null(attr(terms(formula, specials = "strata"), "specials")$strata)
+}
+has_special_arm <- function(formula) {
+  !is.null(attr(terms(formula, specials = "arm"), "arm")$arm)
+}
 
 #' Identity function to indicate specials in formulas
 #'