+++ NEWS.md
@@ -6,6 +6,7 @@
 * Addes `footnotes`, `footnotes<-` and `add_footnotes<-` functions to deal with footnotes
 * Add argument `conf_int` for confidence interval level to `t_el_forest_rps`, `t_forest_rsp`, `t_el_forest_tte``, `t_forest_tte`
 * Add argument `col_symbol_size` to `g_forest` to control the relative size of symbols used in the plot.
+* Added `s_coxph_pairwise` function to perform pairwise testing which is used by `t_tte` and `t_coxph`
 
 # tern 0.6.4